replaces floats with doubles, reduced runtime by 6% by caching CaT current instead of recalcuating it in NMDA_current()

+++ neuron.h
@@ -26,7 +26,7 @@ struct neuron_compartment {
 struct compartment_state {
   int num_params;
   char ** names;
-  float * values;
+  double * values;
 };
 
 struct init_compartment_states {
@@ -37,7 +37,7 @@ struct init_compartment_states {
 struct neuron_params {
   int num_params;
   char ** names;
-  float * values;
+  double * values;
 };
 
 struct network {
@@ -52,8 +52,8 @@ struct neuron_link {
   long from_compartment;
   long to;
   long to_compartment;
-  float weight;
-  float conduction_time;
+  double weight;
+  double conduction_time;
 };
 
 struct link_queue {
@@ -66,8 +66,8 @@ struct link_node {
   long from_compartment;
   long to;
   long to_compartment;
-  float weight;
-  float ctime;
+  double weight;
+  double ctime;
   struct link_node * next;
 };
 
@@ -102,8 +102,8 @@ void cleanup(struct network * network, struct init_compartment_states * init_com
 FILE * open_file_to_section(char * filename, char * section);
 void remove_newline(char * line);
 void link_neurons(struct network * network, char * filename);
-struct neuron_link * create_link(long from, long from_compartment, long to, long to_compartment, float weight, float ctime);
-void queue_link(struct link_queue * link_queue, long from, long from_compartment, long to, long to_compartment, float weight, float ctime);
+struct neuron_link * create_link(long from, long from_compartment, long to, long to_compartment, double weight, double ctime);
+void queue_link(struct link_queue * link_queue, long from, long from_compartment, long to, long to_compartment, double weight, double ctime);
 void create_queued_links(struct network * network, struct link_queue * link_queue);
 void output_state(struct network * network, struct init_compartment_states * states, struct neuron_params * params, char * filename);
 void write_to_file(FILE * fp, char * line);
